Split out stepping over individual axes to find a periodic orbit
The three axes move independently so finding a periodic orbit (loop) in
the solution can be split in finding the orbits for the axes separately
first and then calculating the Lowest Common Multiple of the three. It
is also possible to only check against the initial state, it is the
state that will have to be passed first to go into the repeating
sequence.